requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
tensorflow==2.20.0rc0
matplotlib